requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
flask==2.3.3
openai==1.0.0
lxml>=5.0.0
selenium==4.15.0
python-dotenv==1.0.0
schedule==1.2.0
feedparser==6.0.10
newspaper3k==0.2.8
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
tensorflow>=2.13.0
transformers>=4.30.0
tokenizers>=0.13.0
datasets>=2.14.0
accelerate>=0.20.0
sentencepiece>=0.1.99
safetensors>=0.3.0
wandb>=0.15.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
regex>=2023.5.5
scholarly>=1.7.11
chromedriver-autoinstaller>=0.6.2
arxiv>=1.4.8
pdfplumber>=0.9.0
PyPDF2>=3.0.1
textstat>=0.7.3
nltk>=3.8.1
spacy>=3.7.2
